+++ eth/backend.go
@@ -462,6 +462,9 @@ func (s *Ethereum) APIs() []rpc.API {
 		}, {
 			Namespace: "net",
 			Service:   s.netRPCService,
+		}, {
+			Namespace: "kss",
+			Service:   NewHemiAPI(s),
 		},
 	}...)
 }
