+++ internal/ethapi/backend.go
@@ -129,9 +129,6 @@ func GetAPIs(apiBackend Backend) []rpc.API {
 		}, {
 			Namespace: "personal",
 			Service:   NewPersonalAccountAPI(apiBackend, nonceLock),
-		}, {
-			Namespace: "kss",
-			Service:   NewHemiAPI(apiBackend),
 		},
 	}
 }
